+++ packages/nimble-components/src/rich-text/mention-listbox/tests/mention-listbox.spec.ts
@@ -45,7 +45,8 @@ describe('RichTextMentionListbox', () => {
         await waitAnimationFrame(); // necessary because scrolling is queued with requestAnimationFrame
     }
 
-    it('should scroll the selected option into view when opened', async () => {
+    // Intermittent: see https://github.com/ni/nimble/issues/1891
+    xit('should scroll the selected option into view when opened', async () => {
         const model = new Model();
         const { connect, disconnect } = await setup500Options(model);
         await connect();